2023-autumn-ab-python-hw-14-team-dev-OlgaShish/pyproject.toml
[build-system]
requires = ["setuptools>=68.1"]
build-backend = "setuptools.build_meta"

[tool.poetry]
name = "homework"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.urls]
"Homepage" = "https://github.com/central-university-dev/python-hw-10-template"
"Bug Tracker" = "https://github.com/central-university-dev/python-hw-10-template/issues"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.105.0"
uvicorn = "^0.15.0"
pydantic = "^2.5.3"
pyjwt = "^2.8.0"
httpx = "^0.26.0"
freezegun = "^1.4.0"
sqlalchemy = "^2.0.24"


[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-asyncio = "^0.15.1"
pytest-mock = "^3.12.0"
pytest-cov = "^4.1"
pylint = "^3.0.1"
watchdog = "^3.0.0"
black = "^23.9.1"
flake8 = "^6.1.0"
flake8-todo = "^0.7"
types-requests = "^2.31.0.9"
mypy = "^1.6.0"


[tool.black]
line-length = 79
skip-string-normalization = true